refactored the text snatching to turn it into a oneliner in the Attachment table
